Divide style from index
