+++ src/components/atoms/Button/index.tsx
@@ -1,12 +1,7 @@
 import React from 'react';
-import {
-  makeStyles,
-  createStyles,
-  createMuiTheme,
-  ThemeProvider,
-} from '@material-ui/core/styles';
+import { ThemeProvider } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
-import { primaryColor, secondaryColor } from '../../../utils/color';
+import { theme, useStyles } from './style';
 
 type Props = {
   useCase: UseCase;
@@ -14,32 +9,6 @@ type Props = {
 };
 
 export type UseCase = 'auth' | 'main';
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    auth: {
-      width: '30ch',
-    },
-    main: {
-      width: 100,
-    },
-    button: {
-      marginTop: theme.spacing(1),
-      marginBottom: theme.spacing(1),
-    },
-  })
-);
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      main: primaryColor(),
-    },
-    secondary: {
-      main: secondaryColor(),
-    },
-  },
-});
 
 export default function ContainedButton(props: Props) {
   const { useCase, children } = props;
