+++ src/components/atoms/TextField/index.tsx
@@ -1,5 +1,4 @@
 import React, { useRef } from 'react';
-import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import VisilityIcon from '@material-ui/icons/Visibility';
 import VisilityIconOff from '@material-ui/icons/VisibilityOff';
 import FormControll from '@material-ui/core/FormControl';
@@ -8,7 +7,7 @@ import InputAdornment from '@material-ui/core/InputAdornment';
 import IconButton from '@material-ui/core/IconButton';
 import classNames from 'classnames';
 import { OutlinedInput } from '@material-ui/core';
-import styled from 'styled-components';
+import { TextFieldContainer, useStyles } from './style';
 
 type Props = {
   textFieldType: TextFieldType;
@@ -37,25 +36,6 @@ interface TextField {
   bookPrice: number;
   purchaseDate: string;
 }
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      display: 'flex',
-      flexWrap: 'wrap',
-    },
-    margin: {
-      margin: theme.spacing(1),
-    },
-    textField: {
-      width: '30ch',
-    },
-  })
-);
-
-const TextFieldContainer = styled.div`
-  align-items: center;
-`;
 
 const returnOutlinedtextFieldType = (textFieldType: TextFieldType) => {
   switch (textFieldType) {
