+++ src/components/molecules/BookCard/index.tsx
@@ -1,34 +1,16 @@
 import React from 'react';
-import { makeStyles, createStyles } from '@material-ui/core/styles';
-import styled from 'styled-components';
 import Card from '@material-ui/core/Card';
 import CardMedia from '@material-ui/core/CardMedia';
 import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
 import { Link } from 'react-router-dom';
+import { BookPriceAndDateContainer, useStyles } from './style';
 
 type Props = {
   bookName: string;
   bookPrice?: number;
   purchaseDate?: string;
 };
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    root: {
-      maxWidth: 345,
-    },
-    media: {
-      height: 0,
-      paddingTop: '56.25%',
-    },
-  })
-);
-
-const BookPriceAndDateContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
 
 export default function BookCard(props: Props) {
   const { bookName, bookPrice, purchaseDate } = props;
