+++ src/components/molecules/InputBook/style.ts
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const TextFieldContainer = styled.div`
+  display: flex;
+  flex-flow: column;
+  align-items: center;
+`;
+
+export const Img = styled.img`
+  width: 16rem;
+  height: 16rem;
+`;
+
+export const InputBookContainer = styled.div`
+  display: flex;
+  flex-flow: row;
+  justify-content: space-evenly;
+  align-items: center;
+  margin: 50px 20px;
+`;
