+++ src/components/organisms/Auth/style.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: flex;
+  flex-flow: column;
+  align-items: center;
+`;
