+++ src/components/pages/Login/style.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const Title = styled.h2`
+  text-align: center;
+`;
+
+export const PaperContainer = styled.div`
+  width: 400px;
+  margin: 0 auto;
+`;
